// src/app/services/models/page-response-event-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { EventDto } from '../models/event-dto';
export interface PageResponseEventDto {
  content?: Array<EventDto>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}